Kolejna poprawka ze specjalizacjami

+++ SystemRezerwacjiWizyt/SystemRezerwacjiWizyt/Controllers/SpecializationsController.cs
@@ -38,11 +38,17 @@ namespace SystemRezerwacjiWizyt.Models
             if (Session["User"] is Admin)
             {
                 var sp = db.Specializations.Find(ID);
+                var specs =
+                  db.Specializations.ToList();
+                if (sp.Doctors.Count != 0)
+                {
+                    ViewBag.MSG = "Nie możesz usunąć tej specjalizacji, istnieje doktor który ją ma";
+                    return View("Index", specs);
+                }
                 db.BeginTransaction();
                 db.Specializations.Remove(sp);
                 db.Commit();
-                var specs =
-                   db.Specializations.ToList();
+               
                 return View("Index", specs);
             }
             return RedirectToAction("Index", "Home");
